Move Model to common
